requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
alembic==1.4.2
